Fixes crackTimeDisplay returns "instant" for large passwords

+++ Zxcvbn/Zxcvbn/DBScorer.m
@@ -360,7 +360,7 @@ static int kNumDays = 31;
     return digits + upper + lower + symbols;
 }
 
-- (NSString *)displayTime:(int)seconds
+- (NSString *)displayTime:(float)seconds
 {
     int minute = 60;
     int hour = minute * 60;
